refactor(cli): split shim dispatch into cohesive submodules
shim_dispatch.rs had accreted five separable concerns. Dispatch and
process handoff stay in the module root; candidate identity
fingerprinting, the trust registry and prompt, runtime
materialization, and the Windows node dispatcher move to identity,
trust, runtime_env, and windows submodules. materialize_runtime's
trusted-config resolution and install-config hardening become named
functions carrying their security rationale as docs.
